Crate havok_classes

If the file name is the same as the struct name representing each generated class, a troublesome error will occur in namespace invocation, so a trailing _ is added to avoid this.

e.g. hkClass -> struct hkClass in hkClass_.rs file.

Note

Note that the following enum and bitFlags have duplicate names.

To use it correctly, you must import it from a module as follows.

e.g. havok_classes::hkbEvaluateHandleModifier_::HandleChangeMode

enum HandleChangeMode(TYPE_ENUM)

hkbEvaluateHandleModifier

hkbHandIkControlData

enum Flags(TYPE_FLAGS)

hkbEventInfo

hkMeshVertexBuffer

enum FlagValues(TYPE_FLAGS)

hkClass

hkClassMember

enum Type(TYPE_ENUM)

hkClassMember

hkResourceBase

Modules

BGSGamebryoSequenceGenerator_

BSBoneSwitchGeneratorBoneData_

BSBoneSwitchGenerator_

BSComputeAddBoneAnimModifier_

BSCyclicBlendTransitionGenerator_

BSDecomposeVectorModifier_

BSDirectAtModifier_

BSDistTriggerModifier_

BSEventEveryNEventsModifier_

BSEventOnDeactivateModifier_

BSEventOnFalseToTrueModifier_

BSGetTimeStepModifier_

BSIStateManagerModifierBSIStateManagerStateListener_

BSIStateManagerModifierBSiStateData_

BSIStateManagerModifier_

BSInterpValueModifier_

BSIsActiveModifier_

BSLimbIKModifier_

BSLookAtModifierBoneData_

BSLookAtModifier_

BSModifyOnceModifier_

BSOffsetAnimationGenerator_

BSPassByTargetTriggerModifier_

BSRagdollContactListenerModifier_

BSSpeedSamplerModifier_

BSSynchronizedClipGenerator_

BSTimerModifier_

BSTweenerModifier_

BSiStateTaggingGenerator_

hkAabbHalf_

hkAabbUint32_

hkAabb_

hkAlignSceneToNodeOptions_

hkArrayTypeAttribute_

hkBaseObject_

hkBitField_

hkClassEnumItem_

hkClassEnum_

hkClassMember_

hkClass_

hkColor_

hkContactPointMaterial_

hkContactPoint_

hkCustomAttributesAttribute_

hkCustomAttributes_

hkDataObjectTypeAttribute_

hkDescriptionAttribute_

hkDocumentationAttribute_

hkGeometryTriangle_

hkGeometry_

hkGizmoAttribute_

hkHalf8_

hkIndexedTransformSet_

hkLinkAttribute_

hkLocalFrameGroup_

hkLocalFrame_

hkMemoryMeshBody_

hkMemoryMeshMaterial_

hkMemoryMeshShape_

hkMemoryMeshTexture_

hkMemoryMeshVertexBuffer_

hkMemoryResourceContainer_

hkMemoryResourceHandleExternalLink_

hkMemoryResourceHandle_

hkMemoryTrackerAttribute_

hkMeshBody_

hkMeshBoneIndexMapping_

hkMeshMaterial_

hkMeshSectionCinfo_

hkMeshSection_

hkMeshShape_

hkMeshTexture_

hkMeshVertexBuffer_

hkModelerNodeTypeAttribute_

hkMonitorStreamColorTableColorPair_

hkMonitorStreamColorTable_

hkMonitorStreamFrameInfo_

hkMonitorStreamStringMapStringMap_

hkMonitorStreamStringMap_

hkMoppBvTreeShapeBase_

hkMotionState_

hkMultiThreadCheck_

hkMultipleVertexBufferElementInfo_

hkMultipleVertexBufferLockedElement_

hkMultipleVertexBufferVertexBufferInfo_

hkMultipleVertexBuffer_

hkPackedVector3_

hkPackfileHeader_

hkPackfileSectionHeader_

hkPostFinishAttribute_

hkQTransform_

hkRangeInt32Attribute_

hkRangeRealAttribute_

hkReferencedObject_

hkReflectedFileAttribute_

hkResourceBase_

hkResourceContainer_

hkResourceHandle_

hkRootLevelContainerNamedVariant_

hkRootLevelContainer_

hkSemanticsAttribute_

hkSimpleLocalFrame_

hkSphere_

hkSweptTransform_

hkTraceStreamTitle_

hkTrackerSerializableScanSnapshotAllocation_

hkTrackerSerializableScanSnapshotBlock_

hkTrackerSerializableScanSnapshot_

hkUiAttribute_

hkVertexFormatElement_

hkVertexFormat_

hkWorldMemoryAvailableWatchDog_

hkaAnimatedReferenceFrame_

hkaAnimationBinding_

hkaAnimationContainer_

hkaAnimationPreviewColorContainer_

hkaAnimation_

hkaAnnotationTrackAnnotation_

hkaAnnotationTrack_

hkaBoneAttachment_

hkaBone_

hkaDefaultAnimatedReferenceFrame_

hkaDeltaCompressedAnimationQuantizationFormat_

hkaDeltaCompressedAnimation_

hkaFootstepAnalysisInfoContainer_

hkaFootstepAnalysisInfo_

hkaInterleavedUncompressedAnimation_

hkaKeyFrameHierarchyUtilityControlData_

hkaKeyFrameHierarchyUtility_

hkaMeshBindingMapping_

hkaMeshBinding_

hkaQuantizedAnimationTrackCompressionParams_

hkaQuantizedAnimation_

hkaRagdollInstance_

hkaSkeletonLocalFrameOnBone_

hkaSkeletonMapperDataChainMapping_

hkaSkeletonMapperDataSimpleMapping_

hkaSkeletonMapperData_

hkaSkeletonMapper_

hkaSkeleton_

hkaSplineCompressedAnimationAnimationCompressionParams_

hkaSplineCompressedAnimationTrackCompressionParams_

hkaSplineCompressedAnimation_

hkaWaveletCompressedAnimationCompressionParams_

hkaWaveletCompressedAnimationQuantizationFormat_

hkaWaveletCompressedAnimation_

hkbAttachmentModifier_

hkbAttachmentSetup_

hkbAttributeModifierAssignment_

hkbAttributeModifier_

hkbAuxiliaryNodeInfo_

hkbBehaviorEventsInfo_

hkbBehaviorGraphData_

hkbBehaviorGraphInternalStateInfo_

hkbBehaviorGraphInternalState_

hkbBehaviorGraphStringData_

hkbBehaviorGraph_

hkbBehaviorInfoIdToNamePair_

hkbBehaviorInfo_

hkbBehaviorReferenceGenerator_

hkbBindable_

hkbBlendCurveUtils_

hkbBlenderGeneratorChildInternalState_

hkbBlenderGeneratorChild_

hkbBlenderGeneratorInternalState_

hkbBlenderGenerator_

hkbBlendingTransitionEffectInternalState_

hkbBlendingTransitionEffect_

hkbBoneIndexArray_

hkbBoneWeightArray_

hkbBoolVariableSequencedDataSample_

hkbBoolVariableSequencedData_

hkbCameraShakeEventPayload_

hkbCharacterAddedInfo_

hkbCharacterControlCommand_

hkbCharacterControllerControlData_

hkbCharacterControllerModifierInternalState_

hkbCharacterControllerModifier_

hkbCharacterDataCharacterControllerInfo_

hkbCharacterData_

hkbCharacterInfo_

hkbCharacterSetup_

hkbCharacterSkinInfo_

hkbCharacterSteppedInfo_

hkbCharacterStringData_

hkbCharacter_

hkbClientCharacterState_

hkbClipGeneratorEcho_

hkbClipGeneratorInternalState_

hkbClipGenerator_

hkbClipTriggerArray_

hkbClipTrigger_

hkbCombineTransformsModifierInternalState_

hkbCombineTransformsModifier_

hkbCompiledExpressionSetToken_

hkbCompiledExpressionSet_

hkbComputeDirectionModifierInternalState_

hkbComputeDirectionModifier_

hkbComputeRotationFromAxisAngleModifierInternalState_

hkbComputeRotationFromAxisAngleModifier_

hkbComputeRotationToTargetModifierInternalState_

hkbComputeRotationToTargetModifier_

hkbCondition_

hkbContext_

hkbDampingModifierInternalState_

hkbDampingModifier_

hkbDefaultMessageLog_

hkbDelayedModifierInternalState_

hkbDelayedModifier_

hkbDetectCloseToGroundModifierInternalState_

hkbDetectCloseToGroundModifier_

hkbEvaluateExpressionModifierInternalExpressionData_

hkbEvaluateExpressionModifierInternalState_

hkbEvaluateExpressionModifier_

hkbEvaluateHandleModifier_

hkbEventBase_

hkbEventDrivenModifierInternalState_

hkbEventDrivenModifier_

hkbEventInfo_

hkbEventPayloadList_

hkbEventPayload_

hkbEventProperty_

hkbEventRaisedInfo_

hkbEventRangeDataArray_

hkbEventRangeData_

hkbEventSequencedDataSequencedEvent_

hkbEventSequencedData_

hkbEvent_

hkbEventsFromRangeModifierInternalState_

hkbEventsFromRangeModifier_

hkbExpressionCondition_

hkbExpressionDataArray_

hkbExpressionData_

hkbExtractRagdollPoseModifier_

hkbFootIkControlData_

hkbFootIkControlsModifierLeg_

hkbFootIkControlsModifier_

hkbFootIkDriverInfoLeg_

hkbFootIkDriverInfo_

hkbFootIkGains_

hkbFootIkModifierInternalLegData_

hkbFootIkModifierLeg_

hkbFootIkModifier_

hkbGeneratorOutputListener_

hkbGeneratorSyncInfoSyncPoint_

hkbGeneratorSyncInfo_

hkbGeneratorTransitionEffectInternalState_

hkbGeneratorTransitionEffect_

hkbGenerator_

hkbGetHandleOnBoneModifier_

hkbGetUpModifierInternalState_

hkbGetUpModifier_

hkbGetWorldFromModelModifierInternalState_

hkbGetWorldFromModelModifier_

hkbHandIkControlData_

hkbHandIkControlsModifierHand_

hkbHandIkControlsModifier_

hkbHandIkDriverInfoHand_

hkbHandIkDriverInfo_

hkbHandIkModifierHand_

hkbHandIkModifier_

hkbHandle_

hkbIntEventPayload_

hkbIntVariableSequencedDataSample_

hkbIntVariableSequencedData_

hkbKeyframeBonesModifierKeyframeInfo_

hkbKeyframeBonesModifier_

hkbLinkedSymbolInfo_

hkbLookAtModifierInternalState_

hkbLookAtModifier_

hkbManualSelectorGeneratorInternalState_

hkbManualSelectorGenerator_

hkbMessageLog_

hkbMirrorModifier_

hkbMirroredSkeletonInfo_

hkbModifierGenerator_

hkbModifierList_

hkbModifierWrapper_

hkbModifier_

hkbMoveCharacterModifierInternalState_

hkbMoveCharacterModifier_

hkbNamedEventPayload_

hkbNamedIntEventPayload_

hkbNamedRealEventPayload_

hkbNamedStringEventPayload_

hkbNodeInternalStateInfo_

hkbNode_

hkbParticleSystemEventPayload_

hkbPoseMatchingGeneratorInternalState_

hkbPoseMatchingGenerator_

hkbPoweredRagdollControlData_

hkbPoweredRagdollControlsModifier_

hkbProjectData_

hkbProjectStringData_

hkbProxyModifierProxyInfo_

hkbProxyModifier_

hkbRaiseEventCommand_

hkbRealEventPayload_

hkbRealVariableSequencedDataSample_

hkbRealVariableSequencedData_

hkbReferencePoseGenerator_

hkbRegisteredGenerator_

hkbRigidBodyRagdollControlData_

hkbRigidBodyRagdollControlsModifier_

hkbRoleAttribute_

hkbRotateCharacterModifierInternalState_

hkbRotateCharacterModifier_

hkbSenseHandleModifierRange_

hkbSenseHandleModifier_

hkbSequenceInternalState_

hkbSequenceStringData_

hkbSequence_

hkbSequencedData_

hkbSetBehaviorCommand_

hkbSetLocalTimeOfClipGeneratorCommand_

hkbSetNodePropertyCommand_

hkbSetWordVariableCommand_

hkbSetWorldFromModelModifier_

hkbSimulationControlCommand_

hkbSimulationStateInfo_

hkbStateChooser_

hkbStateListener_

hkbStateMachineActiveTransitionInfo_

hkbStateMachineDelayedTransitionInfo_

hkbStateMachineEventPropertyArray_

hkbStateMachineInternalState_

hkbStateMachineNestedStateMachineData_

hkbStateMachineProspectiveTransitionInfo_

hkbStateMachineStateInfo_

hkbStateMachineTimeInterval_

hkbStateMachineTransitionInfoArray_

hkbStateMachineTransitionInfoReference_

hkbStateMachineTransitionInfo_

hkbStateMachine_

hkbStringCondition_

hkbStringEventPayload_

hkbTestStateChooser_

hkbTimerModifierInternalState_

hkbTimerModifier_

hkbTransformVectorModifierInternalState_

hkbTransformVectorModifier_

hkbTransitionEffect_

hkbTwistModifier_

hkbVariableBindingSetBinding_

hkbVariableBindingSet_

hkbVariableInfo_

hkbVariableValueSet_

hkbVariableValue_

hkbWorldEnums_

hkbWorldFromModelModeData_

hkp2dAngConstraintAtom_

hkpAabbPhantom_

hkpAction_

hkpAgent1nSector_

hkpAngConstraintAtom_

hkpAngFrictionConstraintAtom_

hkpAngLimitConstraintAtom_

hkpAngMotorConstraintAtom_

hkpAngularDashpotAction_

hkpArrayAction_

hkpBallAndSocketConstraintDataAtoms_

hkpBallAndSocketConstraintData_

hkpBallGun_

hkpBallSocketChainDataConstraintInfo_

hkpBallSocketChainData_

hkpBallSocketConstraintAtom_

hkpBinaryAction_

hkpBoxMotion_

hkpBoxShape_

hkpBreakableBody_

hkpBreakableConstraintData_

hkpBridgeAtoms_

hkpBridgeConstraintAtom_

hkpBroadPhaseHandle_

hkpBvShape_

hkpBvTreeShape_

hkpCachingShapePhantom_

hkpCallbackConstraintMotor_

hkpCapsuleShape_

hkpCdBody_

hkpCenterOfMassChangerModifierConstraintAtom_

hkpCharacterControllerCinfo_

hkpCharacterMotion_

hkpCharacterProxyCinfo_

hkpCharacterRigidBodyCinfo_

hkpCogWheelConstraintAtom_

hkpCogWheelConstraintDataAtoms_

hkpCogWheelConstraintData_

hkpCollidableBoundingVolumeData_

hkpCollidableCollidableFilter_

hkpCollidable_

hkpCollisionFilterList_

hkpCollisionFilter_

hkpCompressedMeshShapeBigTriangle_

hkpCompressedMeshShapeChunk_

hkpCompressedMeshShapeConvexPiece_

hkpCompressedMeshShape_

hkpCompressedSampledHeightFieldShape_

hkpConeLimitConstraintAtom_

hkpConstrainedSystemFilter_

hkpConstraintAtom_

hkpConstraintChainData_

hkpConstraintChainInstanceAction_

hkpConstraintChainInstance_

hkpConstraintCollisionFilter_

hkpConstraintData_

hkpConstraintInstanceSmallArraySerializeOverrideType_

hkpConstraintInstance_

hkpConstraintMotor_

hkpConvexListFilter_

hkpConvexListShape_

hkpConvexPieceMeshShape_

hkpConvexPieceStreamData_

hkpConvexShape_

hkpConvexTransformShapeBase_

hkpConvexTransformShape_

hkpConvexTranslateShape_

hkpConvexVerticesConnectivity_

hkpConvexVerticesShapeFourVectors_

hkpConvexVerticesShape_

hkpCylinderShape_

hkpDashpotAction_

hkpDefaultConvexListFilter_

hkpDefaultWorldMemoryWatchDog_

hkpDisableEntityCollisionFilter_

hkpDisplayBindingDataPhysicsSystem_

hkpDisplayBindingDataRigidBody_

hkpDisplayBindingData_

hkpEntityExtendedListeners_

hkpEntitySmallArraySerializeOverrideType_

hkpEntitySpuCollisionCallback_

hkpEntity_

hkpExtendedMeshShapeShapesSubpart_

hkpExtendedMeshShapeSubpart_

hkpExtendedMeshShapeTrianglesSubpart_

hkpExtendedMeshShape_

hkpFastMeshShape_

hkpFirstPersonGun_

hkpFixedRigidMotion_

hkpGenericConstraintDataSchemeConstraintInfo_

hkpGenericConstraintDataScheme_

hkpGenericConstraintData_

hkpGravityGun_

hkpGroupCollisionFilter_

hkpGroupFilter_

hkpHeightFieldShape_

hkpHingeConstraintDataAtoms_

hkpHingeConstraintData_

hkpHingeLimitsDataAtoms_

hkpHingeLimitsData_

hkpIgnoreModifierConstraintAtom_

hkpKeyframedRigidMotion_

hkpLimitedForceConstraintMotor_

hkpLimitedHingeConstraintDataAtoms_

hkpLimitedHingeConstraintData_

hkpLinConstraintAtom_

hkpLinFrictionConstraintAtom_

hkpLinLimitConstraintAtom_

hkpLinMotorConstraintAtom_

hkpLinSoftConstraintAtom_

hkpLinearParametricCurve_

hkpLinkedCollidable_

hkpListShapeChildInfo_

hkpListShape_

hkpMalleableConstraintData_

hkpMassChangerModifierConstraintAtom_

hkpMassProperties_

hkpMaterial_

hkpMaxSizeMotion_

hkpMeshMaterial_

hkpMeshShapeSubpart_

hkpMeshShape_

hkpModifierConstraintAtom_

hkpMoppBvTreeShape_

hkpMoppCodeCodeInfo_

hkpMoppCodeReindexedTerminal_

hkpMoppCode_

hkpMotion_

hkpMotorAction_

hkpMountedBallGun_

hkpMouseSpringAction_

hkpMovingSurfaceModifierConstraintAtom_

hkpMultiRayShapeRay_

hkpMultiRayShape_

hkpMultiSphereShape_

hkpMultithreadedVehicleManager_

hkpNamedMeshMaterial_

hkpNullCollisionFilter_

hkpOverwritePivotConstraintAtom_

hkpPairCollisionFilterMapPairFilterKeyOverrideType_

hkpPairCollisionFilter_

hkpParametricCurve_

hkpPhantomCallbackShape_

hkpPhantom_

hkpPhysicsData_

hkpPhysicsSystemWithContacts_

hkpPhysicsSystem_

hkpPlaneShape_

hkpPointToPathConstraintData_

hkpPointToPlaneConstraintDataAtoms_

hkpPointToPlaneConstraintData_

hkpPositionConstraintMotor_

hkpPoweredChainDataConstraintInfo_

hkpPoweredChainData_

hkpPoweredChainMapperLinkInfo_

hkpPoweredChainMapperTarget_

hkpPoweredChainMapper_

hkpPrismaticConstraintDataAtoms_

hkpPrismaticConstraintData_

hkpProjectileGun_

hkpPropertyValue_

hkpProperty_

hkpPulleyConstraintAtom_

hkpPulleyConstraintDataAtoms_

hkpPulleyConstraintData_

hkpRackAndPinionConstraintAtom_

hkpRackAndPinionConstraintDataAtoms_

hkpRackAndPinionConstraintData_

hkpRagdollConstraintDataAtoms_

hkpRagdollConstraintData_

hkpRagdollLimitsDataAtoms_

hkpRagdollLimitsData_

hkpRagdollMotorConstraintAtom_

hkpRayCollidableFilter_

hkpRayShapeCollectionFilter_

hkpRejectChassisListener_

hkpRemoveTerminalsMoppModifier_

hkpReorientAction_

hkpRigidBody_

hkpRotationalConstraintDataAtoms_

hkpRotationalConstraintData_

hkpSampledHeightFieldShape_

hkpSerializedAgentNnEntry_

hkpSerializedDisplayMarkerList_

hkpSerializedDisplayMarker_

hkpSerializedDisplayRbTransformsDisplayTransformPair_

hkpSerializedDisplayRbTransforms_

hkpSerializedSubTrack1nInfo_

hkpSerializedTrack1nInfo_

hkpSetLocalRotationsConstraintAtom_

hkpSetLocalTransformsConstraintAtom_

hkpSetLocalTranslationsConstraintAtom_

hkpSetupStabilizationAtom_

hkpShapeCollectionFilter_

hkpShapeCollection_

hkpShapeContainer_

hkpShapeInfo_

hkpShapeModifier_

hkpShapePhantom_

hkpShape_

hkpSimpleContactConstraintAtom_

hkpSimpleContactConstraintDataInfo_

hkpSimpleMeshShapeTriangle_

hkpSimpleMeshShape_

hkpSimpleShapePhantomCollisionDetail_

hkpSimpleShapePhantom_

hkpSimulation_

hkpSingleShapeContainer_

hkpSoftContactModifierConstraintAtom_

hkpSphereMotion_

hkpSphereRepShape_

hkpSphereShape_

hkpSpringAction_

hkpSpringDamperConstraintMotor_

hkpStiffSpringChainDataConstraintInfo_

hkpStiffSpringChainData_

hkpStiffSpringConstraintAtom_

hkpStiffSpringConstraintDataAtoms_

hkpStiffSpringConstraintData_

hkpStorageExtendedMeshShapeMaterial_

hkpStorageExtendedMeshShapeMeshSubpartStorage_

hkpStorageExtendedMeshShapeShapeSubpartStorage_

hkpStorageExtendedMeshShape_

hkpStorageMeshShapeSubpartStorage_

hkpStorageMeshShape_

hkpStorageSampledHeightFieldShape_

hkpThinBoxMotion_

hkpTransformShape_

hkpTriSampledHeightFieldBvTreeShape_

hkpTriSampledHeightFieldCollection_

hkpTriangleShape_

hkpTriggerVolumeEventInfo_

hkpTriggerVolume_

hkpTwistLimitConstraintAtom_

hkpTypedBroadPhaseHandle_

hkpTyremarkPoint_

hkpTyremarksInfo_

hkpTyremarksWheel_

hkpUnaryAction_

hkpVehicleAerodynamics_

hkpVehicleBrake_

hkpVehicleCastBatchingManager_

hkpVehicleDataWheelComponentParams_

hkpVehicleData_

hkpVehicleDefaultAerodynamics_

hkpVehicleDefaultAnalogDriverInput_

hkpVehicleDefaultBrakeWheelBrakingProperties_

hkpVehicleDefaultBrake_

hkpVehicleDefaultEngine_

hkpVehicleDefaultSteering_

hkpVehicleDefaultSuspensionWheelSpringSuspensionParameters_

hkpVehicleDefaultSuspension_

hkpVehicleDefaultTransmission_

hkpVehicleDefaultVelocityDamper_

hkpVehicleDriverInputAnalogStatus_

hkpVehicleDriverInputStatus_

hkpVehicleDriverInput_

hkpVehicleEngine_

hkpVehicleFrictionDescriptionAxisDescription_

hkpVehicleFrictionDescription_

hkpVehicleFrictionStatusAxisStatus_

hkpVehicleFrictionStatus_

hkpVehicleInstanceWheelInfo_

hkpVehicleInstance_

hkpVehicleLinearCastBatchingManager_

hkpVehicleLinearCastWheelCollideWheelState_

hkpVehicleLinearCastWheelCollide_

hkpVehicleManager_

hkpVehicleRayCastBatchingManager_

hkpVehicleRayCastWheelCollide_

hkpVehicleSteering_

hkpVehicleSuspensionSuspensionWheelParameters_

hkpVehicleSuspension_

hkpVehicleTransmission_

hkpVehicleVelocityDamper_

hkpVehicleWheelCollide_

hkpVelocityConstraintMotor_

hkpViscousSurfaceModifierConstraintAtom_

hkpWeldingUtility_

hkpWheelConstraintDataAtoms_

hkpWheelConstraintData_

hkpWorldCinfo_

hkpWorldObject_

hkpWorld_

hkxAnimatedFloat_

hkxAnimatedMatrix_

hkxAnimatedQuaternion_

hkxAnimatedVector_

hkxAttributeGroup_

hkxAttributeHolder_

hkxAttribute_

hkxCamera_

hkxEdgeSelectionChannel_

hkxEnumItem_

hkxEnum_

hkxEnvironmentVariable_

hkxEnvironment_

hkxIndexBuffer_

hkxLight_

hkxMaterialEffect_

hkxMaterialProperty_

hkxMaterialShaderSet_

hkxMaterialShader_

hkxMaterialTextureStage_

hkxMaterial_

hkxMeshSection_

hkxMeshUserChannelInfo_

hkxMesh_

hkxNodeAnnotationData_

hkxNodeSelectionSet_

hkxNode_

hkxScene_

hkxSkinBinding_

hkxSparselyAnimatedBool_

hkxSparselyAnimatedEnum_

hkxSparselyAnimatedInt_

hkxSparselyAnimatedString_

hkxTextureFile_

hkxTextureInplace_

hkxTriangleSelectionChannel_

hkxVertexBufferVertexData_

hkxVertexBuffer_

hkxVertexDescriptionElementDecl_

hkxVertexDescription_

hkxVertexFloatDataChannel_

hkxVertexIntDataChannel_

hkxVertexSelectionChannel_

hkxVertexVectorDataChannel_

Structs

BGSGamebryoSequenceGenerator

C++ Info

BSBoneSwitchGenerator

C++ Info

BSBoneSwitchGeneratorBoneData

C++ Info

BSComputeAddBoneAnimModifier

C++ Info

BSCyclicBlendTransitionGenerator

C++ Info

BSDecomposeVectorModifier

C++ Info

BSDirectAtModifier

C++ Info

BSDistTriggerModifier

C++ Info

BSEventEveryNEventsModifier

C++ Info

BSEventOnDeactivateModifier

C++ Info

BSEventOnFalseToTrueModifier

C++ Info

BSGetTimeStepModifier

C++ Info

BSIStateManagerModifier

C++ Info

BSIStateManagerModifierBSIStateManagerStateListener

C++ Info

BSIStateManagerModifierBSiStateData

C++ Info

BSInterpValueModifier

C++ Info

BSIsActiveModifier

C++ Info

BSLimbIKModifier

C++ Info

BSLookAtModifier

C++ Info

BSLookAtModifierBoneData

C++ Info

BSModifyOnceModifier

C++ Info

BSOffsetAnimationGenerator

C++ Info

BSPassByTargetTriggerModifier

C++ Info

BSRagdollContactListenerModifier

C++ Info

BSSpeedSamplerModifier

C++ Info

BSSynchronizedClipGenerator

C++ Info

BSTimerModifier

C++ Info

BSTweenerModifier

C++ Info

BSiStateTaggingGenerator

C++ Info

ExtendedColors

Enum as bit flags

FlagBits

Bit flags

FlagValues

Bit flags

FlagValues

Bit flags

FlagValues

Bit flags

Flags

Bit flags

Flags

Bit flags

HintFlags

Bit flags

RoleFlags

Bit flags

TransitionFlags

Bit flags

hkAabb

C++ Info

hkAabbHalf

C++ Info

hkAabbUint32

C++ Info

hkAlignSceneToNodeOptions

C++ Info

hkArrayTypeAttribute

C++ Info

hkBaseObject

C++ Info

hkBitField

C++ Info

hkClass

C++ Info

hkClassEnum

C++ Info

hkClassEnumItem

C++ Info

hkClassMember

C++ Info

hkColor

C++ Info

hkContactPoint

C++ Info

hkContactPointMaterial

C++ Info

hkCustomAttributes

C++ Info

hkCustomAttributesAttribute

C++ Info

hkDataObjectTypeAttribute

C++ Info

hkDescriptionAttribute

C++ Info

hkDocumentationAttribute

C++ Info

hkGeometry

C++ Info

hkGeometryTriangle

C++ Info

hkGizmoAttribute

C++ Info

hkHalf8

C++ Info

hkIndexedTransformSet

C++ Info

hkLinkAttribute

C++ Info

hkLocalFrame

C++ Info

hkLocalFrameGroup

C++ Info

hkMemoryMeshBody

C++ Info

hkMemoryMeshMaterial

C++ Info

hkMemoryMeshShape

C++ Info

hkMemoryMeshTexture

C++ Info

hkMemoryMeshVertexBuffer

C++ Info

hkMemoryResourceContainer

C++ Info

hkMemoryResourceHandle

C++ Info

hkMemoryResourceHandleExternalLink

C++ Info

hkMemoryTrackerAttribute

C++ Info

hkMeshBody

C++ Info

hkMeshBoneIndexMapping

C++ Info

hkMeshMaterial

C++ Info

hkMeshSection

C++ Info

hkMeshSectionCinfo

C++ Info

hkMeshShape

C++ Info

hkMeshTexture

C++ Info

hkMeshVertexBuffer

C++ Info

hkModelerNodeTypeAttribute

C++ Info

hkMonitorStreamColorTable

C++ Info

hkMonitorStreamColorTableColorPair

C++ Info

hkMonitorStreamFrameInfo

C++ Info

hkMonitorStreamStringMap

C++ Info

hkMonitorStreamStringMapStringMap

C++ Info

hkMoppBvTreeShapeBase

C++ Info

hkMotionState

C++ Info

hkMultiThreadCheck

C++ Info

hkMultipleVertexBuffer

C++ Info

hkMultipleVertexBufferElementInfo

C++ Info

hkMultipleVertexBufferLockedElement

C++ Info

hkMultipleVertexBufferVertexBufferInfo

C++ Info

hkPackedVector3

C++ Info

hkPackfileHeader

C++ Info

hkPackfileSectionHeader

C++ Info

hkPostFinishAttribute

C++ Info

hkQTransform

C++ Info

hkRangeInt32Attribute

C++ Info

hkRangeRealAttribute

C++ Info

hkReferencedObject

C++ Info

hkReflectedFileAttribute

C++ Info

hkResourceBase

C++ Info

hkResourceContainer

C++ Info

hkResourceHandle

C++ Info

hkRootLevelContainer

C++ Info

hkRootLevelContainerNamedVariant

C++ Info

hkSemanticsAttribute

C++ Info

hkSimpleLocalFrame

C++ Info

hkSphere

C++ Info

hkSweptTransform

C++ Info

hkTraceStreamTitle

C++ Info

hkTrackerSerializableScanSnapshot

C++ Info

hkTrackerSerializableScanSnapshotAllocation

C++ Info

hkTrackerSerializableScanSnapshotBlock

C++ Info

hkUiAttribute

C++ Info

hkVertexFormat

C++ Info

hkVertexFormatElement

C++ Info

hkWorldMemoryAvailableWatchDog

C++ Info

hkaAnimatedReferenceFrame

C++ Info

hkaAnimation

C++ Info

hkaAnimationBinding

C++ Info

hkaAnimationContainer

C++ Info

hkaAnimationPreviewColorContainer

C++ Info

hkaAnnotationTrack

C++ Info

hkaAnnotationTrackAnnotation

C++ Info

hkaBone

C++ Info

hkaBoneAttachment

C++ Info

hkaDefaultAnimatedReferenceFrame

C++ Info

hkaDeltaCompressedAnimation

C++ Info

hkaDeltaCompressedAnimationQuantizationFormat

C++ Info

hkaFootstepAnalysisInfo

C++ Info

hkaFootstepAnalysisInfoContainer

C++ Info

hkaInterleavedUncompressedAnimation

C++ Info

hkaKeyFrameHierarchyUtility

C++ Info

hkaKeyFrameHierarchyUtilityControlData

C++ Info

hkaMeshBinding

C++ Info

hkaMeshBindingMapping

C++ Info

hkaQuantizedAnimation

C++ Info

hkaQuantizedAnimationTrackCompressionParams

C++ Info

hkaRagdollInstance

C++ Info

hkaSkeleton

C++ Info

hkaSkeletonLocalFrameOnBone

C++ Info

hkaSkeletonMapper

C++ Info

hkaSkeletonMapperData

C++ Info

hkaSkeletonMapperDataChainMapping

C++ Info

hkaSkeletonMapperDataSimpleMapping

C++ Info

hkaSplineCompressedAnimation

C++ Info

hkaSplineCompressedAnimationAnimationCompressionParams

C++ Info

hkaSplineCompressedAnimationTrackCompressionParams

C++ Info

hkaWaveletCompressedAnimation

C++ Info

hkaWaveletCompressedAnimationCompressionParams

C++ Info

hkaWaveletCompressedAnimationQuantizationFormat

C++ Info

hkbAttachmentModifier

C++ Info

hkbAttachmentSetup

C++ Info

hkbAttributeModifier

C++ Info

hkbAttributeModifierAssignment

C++ Info

hkbAuxiliaryNodeInfo

C++ Info

hkbBehaviorEventsInfo

C++ Info

hkbBehaviorGraph

C++ Info

hkbBehaviorGraphData

C++ Info

hkbBehaviorGraphInternalState

C++ Info

hkbBehaviorGraphInternalStateInfo

C++ Info

hkbBehaviorGraphStringData

C++ Info

hkbBehaviorInfo

C++ Info

hkbBehaviorInfoIdToNamePair

C++ Info

hkbBehaviorReferenceGenerator

C++ Info

hkbBindable

C++ Info

hkbBlendCurveUtils

C++ Info

hkbBlenderGenerator

C++ Info

hkbBlenderGeneratorChild

C++ Info

hkbBlenderGeneratorChildInternalState

C++ Info

hkbBlenderGeneratorInternalState

C++ Info

hkbBlendingTransitionEffect

C++ Info

hkbBlendingTransitionEffectInternalState

C++ Info

hkbBoneIndexArray

C++ Info

hkbBoneWeightArray

C++ Info

hkbBoolVariableSequencedData

C++ Info

hkbBoolVariableSequencedDataSample

C++ Info

hkbCameraShakeEventPayload

C++ Info

hkbCharacter

C++ Info

hkbCharacterAddedInfo

C++ Info

hkbCharacterControlCommand

C++ Info

hkbCharacterControllerControlData

C++ Info

hkbCharacterControllerModifier

C++ Info

hkbCharacterControllerModifierInternalState

C++ Info

hkbCharacterData

C++ Info

hkbCharacterDataCharacterControllerInfo

C++ Info

hkbCharacterInfo

C++ Info

hkbCharacterSetup

C++ Info

hkbCharacterSkinInfo

C++ Info

hkbCharacterSteppedInfo

C++ Info

hkbCharacterStringData

C++ Info

hkbClientCharacterState

C++ Info

hkbClipGenerator

C++ Info

hkbClipGeneratorEcho

C++ Info

hkbClipGeneratorInternalState

C++ Info

hkbClipTrigger

C++ Info

hkbClipTriggerArray

C++ Info

hkbCombineTransformsModifier

C++ Info

hkbCombineTransformsModifierInternalState

C++ Info

hkbCompiledExpressionSet

C++ Info

hkbCompiledExpressionSetToken

C++ Info

hkbComputeDirectionModifier

C++ Info

hkbComputeDirectionModifierInternalState

C++ Info

hkbComputeRotationFromAxisAngleModifier

C++ Info

hkbComputeRotationFromAxisAngleModifierInternalState

C++ Info

hkbComputeRotationToTargetModifier

C++ Info

hkbComputeRotationToTargetModifierInternalState

C++ Info

hkbCondition

C++ Info

hkbContext

C++ Info

hkbDampingModifier

C++ Info

hkbDampingModifierInternalState

C++ Info

hkbDefaultMessageLog

C++ Info

hkbDelayedModifier

C++ Info

hkbDelayedModifierInternalState

C++ Info

hkbDetectCloseToGroundModifier

C++ Info

hkbDetectCloseToGroundModifierInternalState

C++ Info

hkbEvaluateExpressionModifier

C++ Info

hkbEvaluateExpressionModifierInternalExpressionData

C++ Info

hkbEvaluateExpressionModifierInternalState

C++ Info

hkbEvaluateHandleModifier

C++ Info

hkbEvent

C++ Info

hkbEventBase

C++ Info

hkbEventDrivenModifier

C++ Info

hkbEventDrivenModifierInternalState

C++ Info

hkbEventInfo

C++ Info

hkbEventPayload

C++ Info

hkbEventPayloadList

C++ Info

hkbEventProperty

C++ Info

hkbEventRaisedInfo

C++ Info

hkbEventRangeData

C++ Info

hkbEventRangeDataArray

C++ Info

hkbEventSequencedData

C++ Info

hkbEventSequencedDataSequencedEvent

C++ Info

hkbEventsFromRangeModifier

C++ Info

hkbEventsFromRangeModifierInternalState

C++ Info

hkbExpressionCondition

C++ Info

hkbExpressionData

C++ Info

hkbExpressionDataArray

C++ Info

hkbExtractRagdollPoseModifier

C++ Info

hkbFootIkControlData

C++ Info

hkbFootIkControlsModifier

C++ Info

hkbFootIkControlsModifierLeg

C++ Info

hkbFootIkDriverInfo

C++ Info

hkbFootIkDriverInfoLeg

C++ Info

hkbFootIkGains

C++ Info

hkbFootIkModifier

C++ Info

hkbFootIkModifierInternalLegData

C++ Info

hkbFootIkModifierLeg

C++ Info

hkbGenerator

C++ Info

hkbGeneratorOutputListener

C++ Info

hkbGeneratorSyncInfo

C++ Info

hkbGeneratorSyncInfoSyncPoint

C++ Info

hkbGeneratorTransitionEffect

C++ Info

hkbGeneratorTransitionEffectInternalState

C++ Info

hkbGetHandleOnBoneModifier

C++ Info

hkbGetUpModifier

C++ Info

hkbGetUpModifierInternalState

C++ Info

hkbGetWorldFromModelModifier

C++ Info

hkbGetWorldFromModelModifierInternalState

C++ Info

hkbHandIkControlData

C++ Info

hkbHandIkControlsModifier

C++ Info

hkbHandIkControlsModifierHand

C++ Info

hkbHandIkDriverInfo

C++ Info

hkbHandIkDriverInfoHand

C++ Info

hkbHandIkModifier

C++ Info

hkbHandIkModifierHand

C++ Info

hkbHandle

C++ Info

hkbIntEventPayload

C++ Info

hkbIntVariableSequencedData

C++ Info

hkbIntVariableSequencedDataSample

C++ Info

hkbKeyframeBonesModifier

C++ Info

hkbKeyframeBonesModifierKeyframeInfo

C++ Info

hkbLinkedSymbolInfo

C++ Info

hkbLookAtModifier

C++ Info

hkbLookAtModifierInternalState

C++ Info

hkbManualSelectorGenerator

C++ Info

hkbManualSelectorGeneratorInternalState

C++ Info

hkbMessageLog

C++ Info

hkbMirrorModifier

C++ Info

hkbMirroredSkeletonInfo

C++ Info

hkbModifier

C++ Info

hkbModifierGenerator

C++ Info

hkbModifierList

C++ Info

hkbModifierWrapper

C++ Info

hkbMoveCharacterModifier

C++ Info

hkbMoveCharacterModifierInternalState

C++ Info

hkbNamedEventPayload

C++ Info

hkbNamedIntEventPayload

C++ Info

hkbNamedRealEventPayload

C++ Info

hkbNamedStringEventPayload

C++ Info

hkbNode

C++ Info

hkbNodeInternalStateInfo

C++ Info

hkbParticleSystemEventPayload

C++ Info

hkbPoseMatchingGenerator

C++ Info

hkbPoseMatchingGeneratorInternalState

C++ Info

hkbPoweredRagdollControlData

C++ Info

hkbPoweredRagdollControlsModifier

C++ Info

hkbProjectData

C++ Info

hkbProjectStringData

C++ Info

hkbProxyModifier

C++ Info

hkbProxyModifierProxyInfo

C++ Info

hkbRaiseEventCommand

C++ Info

hkbRealEventPayload

C++ Info

hkbRealVariableSequencedData

C++ Info

hkbRealVariableSequencedDataSample

C++ Info

hkbReferencePoseGenerator

C++ Info

hkbRegisteredGenerator

C++ Info

hkbRigidBodyRagdollControlData

C++ Info

hkbRigidBodyRagdollControlsModifier

C++ Info

hkbRoleAttribute

C++ Info

hkbRotateCharacterModifier

C++ Info

hkbRotateCharacterModifierInternalState

C++ Info

hkbSenseHandleModifier

C++ Info

hkbSenseHandleModifierRange

C++ Info

hkbSequence

C++ Info

hkbSequenceInternalState

C++ Info

hkbSequenceStringData

C++ Info

hkbSequencedData

C++ Info

hkbSetBehaviorCommand

C++ Info

hkbSetLocalTimeOfClipGeneratorCommand

C++ Info

hkbSetNodePropertyCommand

C++ Info

hkbSetWordVariableCommand

C++ Info

hkbSetWorldFromModelModifier

C++ Info

hkbSimulationControlCommand

C++ Info

hkbSimulationStateInfo

C++ Info

hkbStateChooser

C++ Info

hkbStateListener

C++ Info

hkbStateMachine

C++ Info

hkbStateMachineActiveTransitionInfo

C++ Info

hkbStateMachineDelayedTransitionInfo

C++ Info

hkbStateMachineEventPropertyArray

C++ Info

hkbStateMachineInternalState

C++ Info

hkbStateMachineNestedStateMachineData

C++ Info

hkbStateMachineProspectiveTransitionInfo

C++ Info

hkbStateMachineStateInfo

C++ Info

hkbStateMachineTimeInterval

C++ Info

hkbStateMachineTransitionInfo

C++ Info

hkbStateMachineTransitionInfoArray

C++ Info

hkbStateMachineTransitionInfoReference

C++ Info

hkbStringCondition

C++ Info

hkbStringEventPayload

C++ Info

hkbTestStateChooser

C++ Info

hkbTimerModifier

C++ Info

hkbTimerModifierInternalState

C++ Info

hkbTransformVectorModifier

C++ Info

hkbTransformVectorModifierInternalState

C++ Info

hkbTransitionEffect

C++ Info

hkbTwistModifier

C++ Info

hkbVariableBindingSet

C++ Info

hkbVariableBindingSetBinding

C++ Info

hkbVariableInfo

C++ Info

hkbVariableValue

C++ Info

hkbVariableValueSet

C++ Info

hkbWorldEnums

C++ Info

hkbWorldFromModelModeData

C++ Info

hkp2dAngConstraintAtom

C++ Info

hkpAabbPhantom

C++ Info

hkpAction

C++ Info

hkpAgent1nSector

C++ Info

hkpAngConstraintAtom

C++ Info

hkpAngFrictionConstraintAtom

C++ Info

hkpAngLimitConstraintAtom

C++ Info

hkpAngMotorConstraintAtom

C++ Info

hkpAngularDashpotAction

C++ Info

hkpArrayAction

C++ Info

hkpBallAndSocketConstraintData

C++ Info

hkpBallAndSocketConstraintDataAtoms

C++ Info

hkpBallGun

C++ Info

hkpBallSocketChainData

C++ Info

hkpBallSocketChainDataConstraintInfo

C++ Info

hkpBallSocketConstraintAtom

C++ Info

hkpBinaryAction

C++ Info

hkpBoxMotion

C++ Info

hkpBoxShape

C++ Info

hkpBreakableBody

C++ Info

hkpBreakableConstraintData

C++ Info

hkpBridgeAtoms

C++ Info

hkpBridgeConstraintAtom

C++ Info

hkpBroadPhaseHandle

C++ Info

hkpBvShape

C++ Info

hkpBvTreeShape

C++ Info

hkpCachingShapePhantom

C++ Info

hkpCallbackConstraintMotor

C++ Info

hkpCapsuleShape

C++ Info

hkpCdBody

C++ Info

hkpCenterOfMassChangerModifierConstraintAtom

C++ Info

hkpCharacterControllerCinfo

C++ Info

hkpCharacterMotion

C++ Info

hkpCharacterProxyCinfo

C++ Info

hkpCharacterRigidBodyCinfo

C++ Info

hkpCogWheelConstraintAtom

C++ Info

hkpCogWheelConstraintData

C++ Info

hkpCogWheelConstraintDataAtoms

C++ Info

hkpCollidable

C++ Info

hkpCollidableBoundingVolumeData

C++ Info

hkpCollidableCollidableFilter

C++ Info

hkpCollisionFilter

C++ Info

hkpCollisionFilterList

C++ Info

hkpCompressedMeshShape

C++ Info

hkpCompressedMeshShapeBigTriangle

C++ Info

hkpCompressedMeshShapeChunk

C++ Info

hkpCompressedMeshShapeConvexPiece

C++ Info

hkpCompressedSampledHeightFieldShape

C++ Info

hkpConeLimitConstraintAtom

C++ Info

hkpConstrainedSystemFilter

C++ Info

hkpConstraintAtom

C++ Info

hkpConstraintChainData

C++ Info

hkpConstraintChainInstance

C++ Info

hkpConstraintChainInstanceAction

C++ Info

hkpConstraintCollisionFilter

C++ Info

hkpConstraintData

C++ Info

hkpConstraintInstance

C++ Info

hkpConstraintInstanceSmallArraySerializeOverrideType

C++ Info

hkpConstraintMotor

C++ Info

hkpConvexListFilter

C++ Info

hkpConvexListShape

C++ Info

hkpConvexPieceMeshShape

C++ Info

hkpConvexPieceStreamData

C++ Info

hkpConvexShape

C++ Info

hkpConvexTransformShape

C++ Info

hkpConvexTransformShapeBase

C++ Info

hkpConvexTranslateShape

C++ Info

hkpConvexVerticesConnectivity

C++ Info

hkpConvexVerticesShape

C++ Info

hkpConvexVerticesShapeFourVectors

C++ Info

hkpCylinderShape

C++ Info

hkpDashpotAction

C++ Info

hkpDefaultConvexListFilter

C++ Info

hkpDefaultWorldMemoryWatchDog

C++ Info

hkpDisableEntityCollisionFilter

C++ Info

hkpDisplayBindingData

C++ Info

hkpDisplayBindingDataPhysicsSystem

C++ Info

hkpDisplayBindingDataRigidBody

C++ Info

hkpEntity

C++ Info

hkpEntityExtendedListeners

C++ Info

hkpEntitySmallArraySerializeOverrideType

C++ Info

hkpEntitySpuCollisionCallback

C++ Info

hkpExtendedMeshShape

C++ Info

hkpExtendedMeshShapeShapesSubpart

C++ Info

hkpExtendedMeshShapeSubpart

C++ Info

hkpExtendedMeshShapeTrianglesSubpart

C++ Info

hkpFastMeshShape

C++ Info

hkpFirstPersonGun

C++ Info

hkpFixedRigidMotion

C++ Info

hkpGenericConstraintData

C++ Info

hkpGenericConstraintDataScheme

C++ Info

hkpGenericConstraintDataSchemeConstraintInfo

C++ Info

hkpGravityGun

C++ Info

hkpGroupCollisionFilter

C++ Info

hkpGroupFilter

C++ Info

hkpHeightFieldShape

C++ Info

hkpHingeConstraintData

C++ Info

hkpHingeConstraintDataAtoms

C++ Info

hkpHingeLimitsData

C++ Info

hkpHingeLimitsDataAtoms

C++ Info

hkpIgnoreModifierConstraintAtom

C++ Info

hkpKeyframedRigidMotion

C++ Info

hkpLimitedForceConstraintMotor

C++ Info

hkpLimitedHingeConstraintData

C++ Info

hkpLimitedHingeConstraintDataAtoms

C++ Info

hkpLinConstraintAtom

C++ Info

hkpLinFrictionConstraintAtom

C++ Info

hkpLinLimitConstraintAtom

C++ Info

hkpLinMotorConstraintAtom

C++ Info

hkpLinSoftConstraintAtom

C++ Info

hkpLinearParametricCurve

C++ Info

hkpLinkedCollidable

C++ Info

hkpListShape

C++ Info

hkpListShapeChildInfo

C++ Info

hkpMalleableConstraintData

C++ Info

hkpMassChangerModifierConstraintAtom

C++ Info

hkpMassProperties

C++ Info

hkpMaterial

C++ Info

hkpMaxSizeMotion

C++ Info

hkpMeshMaterial

C++ Info

hkpMeshShape

C++ Info

hkpMeshShapeSubpart

C++ Info

hkpModifierConstraintAtom

C++ Info

hkpMoppBvTreeShape

C++ Info

hkpMoppCode

C++ Info

hkpMoppCodeCodeInfo

C++ Info

hkpMoppCodeReindexedTerminal

C++ Info

hkpMotion

C++ Info

hkpMotorAction

C++ Info

hkpMountedBallGun

C++ Info

hkpMouseSpringAction

C++ Info

hkpMovingSurfaceModifierConstraintAtom

C++ Info

hkpMultiRayShape

C++ Info

hkpMultiRayShapeRay

C++ Info

hkpMultiSphereShape

C++ Info

hkpMultithreadedVehicleManager

C++ Info

hkpNamedMeshMaterial

C++ Info

hkpNullCollisionFilter

C++ Info

hkpOverwritePivotConstraintAtom

C++ Info

hkpPairCollisionFilter

C++ Info

hkpPairCollisionFilterMapPairFilterKeyOverrideType

C++ Info

hkpParametricCurve

C++ Info

hkpPhantom

C++ Info

hkpPhantomCallbackShape

C++ Info

hkpPhysicsData

C++ Info

hkpPhysicsSystem

C++ Info

hkpPhysicsSystemWithContacts

C++ Info

hkpPlaneShape

C++ Info

hkpPointToPathConstraintData

C++ Info

hkpPointToPlaneConstraintData

C++ Info

hkpPointToPlaneConstraintDataAtoms

C++ Info

hkpPositionConstraintMotor

C++ Info

hkpPoweredChainData

C++ Info

hkpPoweredChainDataConstraintInfo

C++ Info

hkpPoweredChainMapper

C++ Info

hkpPoweredChainMapperLinkInfo

C++ Info

hkpPoweredChainMapperTarget

C++ Info

hkpPrismaticConstraintData

C++ Info

hkpPrismaticConstraintDataAtoms

C++ Info

hkpProjectileGun

C++ Info

hkpProperty

C++ Info

hkpPropertyValue

C++ Info

hkpPulleyConstraintAtom

C++ Info

hkpPulleyConstraintData

C++ Info

hkpPulleyConstraintDataAtoms

C++ Info

hkpRackAndPinionConstraintAtom

C++ Info

hkpRackAndPinionConstraintData

C++ Info

hkpRackAndPinionConstraintDataAtoms

C++ Info

hkpRagdollConstraintData

C++ Info

hkpRagdollConstraintDataAtoms

C++ Info

hkpRagdollLimitsData

C++ Info

hkpRagdollLimitsDataAtoms

C++ Info

hkpRagdollMotorConstraintAtom

C++ Info

hkpRayCollidableFilter

C++ Info

hkpRayShapeCollectionFilter

C++ Info

hkpRejectChassisListener

C++ Info

hkpRemoveTerminalsMoppModifier

C++ Info

hkpReorientAction

C++ Info

hkpRigidBody

C++ Info

hkpRotationalConstraintData

C++ Info

hkpRotationalConstraintDataAtoms

C++ Info

hkpSampledHeightFieldShape

C++ Info

hkpSerializedAgentNnEntry

C++ Info

hkpSerializedDisplayMarker

C++ Info

hkpSerializedDisplayMarkerList

C++ Info

hkpSerializedDisplayRbTransforms

C++ Info

hkpSerializedDisplayRbTransformsDisplayTransformPair

C++ Info

hkpSerializedSubTrack1nInfo

C++ Info

hkpSerializedTrack1nInfo

C++ Info

hkpSetLocalRotationsConstraintAtom

C++ Info

hkpSetLocalTransformsConstraintAtom

C++ Info

hkpSetLocalTranslationsConstraintAtom

C++ Info

hkpSetupStabilizationAtom

C++ Info

hkpShape

C++ Info

hkpShapeCollection

C++ Info

hkpShapeCollectionFilter

C++ Info

hkpShapeContainer

C++ Info

hkpShapeInfo

C++ Info

hkpShapeModifier

C++ Info

hkpShapePhantom

C++ Info

hkpSimpleContactConstraintAtom

C++ Info

hkpSimpleContactConstraintDataInfo

C++ Info

hkpSimpleMeshShape

C++ Info

hkpSimpleMeshShapeTriangle

C++ Info

hkpSimpleShapePhantom

C++ Info

hkpSimpleShapePhantomCollisionDetail

C++ Info

hkpSimulation

C++ Info

hkpSingleShapeContainer

C++ Info

hkpSoftContactModifierConstraintAtom

C++ Info

hkpSphereMotion

C++ Info

hkpSphereRepShape

C++ Info

hkpSphereShape

C++ Info

hkpSpringAction

C++ Info

hkpSpringDamperConstraintMotor

C++ Info

hkpStiffSpringChainData

C++ Info

hkpStiffSpringChainDataConstraintInfo

C++ Info

hkpStiffSpringConstraintAtom

C++ Info

hkpStiffSpringConstraintData

C++ Info

hkpStiffSpringConstraintDataAtoms

C++ Info

hkpStorageExtendedMeshShape

C++ Info

hkpStorageExtendedMeshShapeMaterial

C++ Info

hkpStorageExtendedMeshShapeMeshSubpartStorage

C++ Info

hkpStorageExtendedMeshShapeShapeSubpartStorage

C++ Info

hkpStorageMeshShape

C++ Info

hkpStorageMeshShapeSubpartStorage

C++ Info

hkpStorageSampledHeightFieldShape

C++ Info

hkpThinBoxMotion

C++ Info

hkpTransformShape

C++ Info

hkpTriSampledHeightFieldBvTreeShape

C++ Info

hkpTriSampledHeightFieldCollection

C++ Info

hkpTriangleShape

C++ Info

hkpTriggerVolume

C++ Info

hkpTriggerVolumeEventInfo

C++ Info

hkpTwistLimitConstraintAtom

C++ Info

hkpTypedBroadPhaseHandle

C++ Info

hkpTyremarkPoint

C++ Info

hkpTyremarksInfo

C++ Info

hkpTyremarksWheel

C++ Info

hkpUnaryAction

C++ Info

hkpVehicleAerodynamics

C++ Info

hkpVehicleBrake

C++ Info

hkpVehicleCastBatchingManager

C++ Info

hkpVehicleData

C++ Info

hkpVehicleDataWheelComponentParams

C++ Info

hkpVehicleDefaultAerodynamics

C++ Info

hkpVehicleDefaultAnalogDriverInput

C++ Info

hkpVehicleDefaultBrake

C++ Info

hkpVehicleDefaultBrakeWheelBrakingProperties

C++ Info

hkpVehicleDefaultEngine

C++ Info

hkpVehicleDefaultSteering

C++ Info

hkpVehicleDefaultSuspension

C++ Info

hkpVehicleDefaultSuspensionWheelSpringSuspensionParameters

C++ Info

hkpVehicleDefaultTransmission

C++ Info

hkpVehicleDefaultVelocityDamper

C++ Info

hkpVehicleDriverInput

C++ Info

hkpVehicleDriverInputAnalogStatus

C++ Info

hkpVehicleDriverInputStatus

C++ Info

hkpVehicleEngine

C++ Info

hkpVehicleFrictionDescription

C++ Info

hkpVehicleFrictionDescriptionAxisDescription

C++ Info

hkpVehicleFrictionStatus

C++ Info

hkpVehicleFrictionStatusAxisStatus

C++ Info

hkpVehicleInstance

C++ Info

hkpVehicleInstanceWheelInfo

C++ Info

hkpVehicleLinearCastBatchingManager

C++ Info

hkpVehicleLinearCastWheelCollide

C++ Info

hkpVehicleLinearCastWheelCollideWheelState

C++ Info

hkpVehicleManager

C++ Info

hkpVehicleRayCastBatchingManager

C++ Info

hkpVehicleRayCastWheelCollide

C++ Info

hkpVehicleSteering

C++ Info

hkpVehicleSuspension

C++ Info

hkpVehicleSuspensionSuspensionWheelParameters

C++ Info

hkpVehicleTransmission

C++ Info

hkpVehicleVelocityDamper

C++ Info

hkpVehicleWheelCollide

C++ Info

hkpVelocityConstraintMotor

C++ Info

hkpViscousSurfaceModifierConstraintAtom

C++ Info

hkpWeldingUtility

C++ Info

hkpWheelConstraintData

C++ Info

hkpWheelConstraintDataAtoms

C++ Info

hkpWorld

C++ Info

hkpWorldCinfo

C++ Info

hkpWorldObject

C++ Info

hkxAnimatedFloat

C++ Info

hkxAnimatedMatrix

C++ Info

hkxAnimatedQuaternion

C++ Info

hkxAnimatedVector

C++ Info

hkxAttribute

C++ Info

hkxAttributeGroup

C++ Info

hkxAttributeHolder

C++ Info

hkxCamera

C++ Info

hkxEdgeSelectionChannel

C++ Info

hkxEnum

C++ Info

hkxEnumItem

C++ Info

hkxEnvironment

C++ Info

hkxEnvironmentVariable

C++ Info

hkxIndexBuffer

C++ Info

hkxLight

C++ Info

hkxMaterial

C++ Info

hkxMaterialEffect

C++ Info

hkxMaterialProperty

C++ Info

hkxMaterialShader

C++ Info

hkxMaterialShaderSet

C++ Info

hkxMaterialTextureStage

C++ Info

hkxMesh

C++ Info

hkxMeshSection

C++ Info

hkxMeshUserChannelInfo

C++ Info

hkxNode

C++ Info

hkxNodeAnnotationData

C++ Info

hkxNodeSelectionSet

C++ Info

hkxScene

C++ Info

hkxSkinBinding

C++ Info

hkxSparselyAnimatedBool

C++ Info

hkxSparselyAnimatedEnum

C++ Info

hkxSparselyAnimatedInt

C++ Info

hkxSparselyAnimatedString

C++ Info

hkxTextureFile

C++ Info

hkxTextureInplace

C++ Info

hkxTriangleSelectionChannel

C++ Info

hkxVertexBuffer

C++ Info

hkxVertexBufferVertexData

C++ Info

hkxVertexDescription

C++ Info

hkxVertexDescriptionElementDecl

C++ Info

hkxVertexFloatDataChannel

C++ Info

hkxVertexIntDataChannel

C++ Info

hkxVertexSelectionChannel

C++ Info

hkxVertexVectorDataChannel

C++ Info

Enums

AbsoluteTimeCounter

C++ Info

AlignMode

C++ Info

AnimationType

C++ Info

ArrayType

C++ Info

AtomType

C++ Info

AttachmentType

C++ Info

BindingType

C++ Info

BlendCurve

C++ Info

BlendHint

C++ Info

BlendModeFunction

C++ Info

BroadPhaseBorderBehaviour

C++ Info

BuildType

C++ Info

BvTreeType

C++ Info

CallbackType

C++ Info

CharacterControlCommand

C++ Info

Classes

CollectionType

C++ Info

ComponentType

C++ Info

ComponentUsage

C++ Info

ConstraintPriority

C++ Info

ContactPointGeneration

C++ Info

DataType

C++ Info

DataUsage

C++ Info

EffectType

C++ Info

EndMode

C++ Info

Event

C++ Info

EventMode

C++ Info

EventRangeMode

C++ Info

ExpressionEventMode

C++ Info

FilterMode

C++ Info

Format

C++ Info

GizmoType

C++ Info

HandleChangeMode

C++ Info

HandleChangeMode

C++ Info

HeightFieldType

C++ Info

HideInModeler

C++ Info

Hint

C++ Info

IndexStridingType

C++ Info

IndexType

C++ Info

InitialVelocityCoordinates

C++ Info

KeyboardKey

C++ Info

LastProcessingStep

C++ Info

LightType

C++ Info

LinearVelocityMode

C++ Info

Link

C++ Info

MappingType

C++ Info

MaterialIndexStridingType

C++ Info

MaterialType

C++ Info

MeasurementMode

C++ Info

MeshSectionIndexType

C++ Info

MeshShapeIndexStridingType

C++ Info

MeshShapeMaterialIndexStridingType

C++ Info

Mode

C++ Info

ModelerType

C++ Info

MotionMode

C++ Info

MotionType

C++ Info

MotorType

C++ Info

NodeType

C++ Info

OnDestructionRemapInfo

C++ Info

Operation

C++ Info

Operator

C++ Info

OrientationConstraintType

C++ Info

PhantomType

C++ Info

PlaybackMode

C++ Info

PrimitiveType

C++ Info

ResponseType

C++ Info

Role

C++ Info

RotationAxisCoordinates

C++ Info

RotationQuantization

C++ Info

ScalarQuantization

C++ Info

SelfTransitionMode

C++ Info

Semantics

C++ Info

SensingMode

C++ Info

SerializedAgentType

C++ Info

SetAngleMethod

C++ Info

ShaderType

C++ Info

SimulationControlCommand

C++ Info

SimulationState

C++ Info

SimulationType

C++ Info

SolvingMethod

C++ Info

Stage

C++ Info

StartStateMode

C++ Info

StateMachineSelfTransitionMode

C++ Info

SubpartType

C++ Info

SystemType

C++ Info

TextureType

C++ Info

TextureUsageType

C++ Info

ToGeneratorState

C++ Info

TokenType

C++ Info

ToolNodeType

C++ Info

TreeUpdateType

C++ Info

Type

C++ Info

Type

C++ Info

Type

C++ Info

Type

C++ Info

VariableMode

C++ Info

VariableType

C++ Info

VertexFloatDimensions

C++ Info

WeldingType

C++ Info

WorldFromModelMode

C++ Info

hkpFilterType

C++ Info