havok_classes::hkColor_

Struct ExtendedColors

pub struct ExtendedColors(/* private fields */);

Enum as bit flags

C++ Info

name: ExtendedColors(ctype: hkEnum<ExtendedColors, hkUint32>)

Why this enum defined as bitflags?

Since Rust does not allow the definition of enum with duplicate discriminant values, they are defined as bitflags.

Implementations

impl ExtendedColors

pub const MAROON: Self

4286578688

pub const DARKRED: Self

4287299584

pub const RED: Self

4294901760

pub const LIGHTPINK: Self

4294944449

pub const CRIMSON: Self

4292613180

pub const PALEVIOLETRED: Self

4292789163

pub const HOTPINK: Self

4294927108

pub const DEEPPINK: Self

4294907027

pub const MEDIUMVIOLETRED: Self

4291237253

pub const PURPLE: Self

4286578816

pub const DARKMAGENTA: Self

4287299723

pub const ORCHID: Self

4292243670

pub const THISTLE: Self

4292394968

pub const PLUM: Self

4292714717

pub const VIOLET: Self

4293821166

pub const FUCHSIA: Self

4294902015

pub const MAGENTA: Self

4294902015

pub const MEDIUMORCHID: Self

4290414451

pub const DARKVIOLET: Self

4287889619

pub const DARKORCHID: Self

4288230092

pub const BLUEVIOLET: Self

4287245282

pub const INDIGO: Self

4284875570

pub const MEDIUMPURPLE: Self

4289454203

pub const SLATEBLUE: Self

4287317261

pub const MEDIUMSLATEBLUE: Self

4288269460

pub const DARKBLUE: Self

4287299587

pub const MEDIUMBLUE: Self

4289379277

pub const BLUE: Self

4294901760

pub const NAVY: Self

4284875648

pub const MIDNIGHTBLUE: Self

4281368880

pub const DARKSLATEBLUE: Self

4284048699

pub const ROYALBLUE: Self

4286006955

pub const CORNFLOWERBLUE: Self

4284782061

pub const LIGHTSTEELBLUE: Self

4289775366

pub const ALICEBLUE: Self

4293982463

pub const GHOSTWHITE: Self

4294638335

pub const LAVENDER: Self

4293322490

pub const DODGERBLUE: Self

4280193279

pub const STEELBLUE: Self

4282811060

pub const DEEPSKYBLUE: Self

4278255615

pub const SLATEGRAY: Self

4285835392

pub const LIGHTSLATEGRAY: Self

4286023833

pub const LIGHTSKYBLUE: Self

4289339194

pub const SKYBLUE: Self

4289342075

pub const LIGHTBLUE: Self

4289959174

pub const TEAL: Self

4284829576

pub const DARKCYAN: Self

4287358955

pub const DARKTURQUOISE: Self

4289470265

pub const CYAN: Self

4286582015

pub const MEDIUMTURQUOISE: Self

4289789324

pub const CADETBLUE: Self

4288893830

pub const PALETURQUOISE: Self

4289520374

pub const LIGHTCYAN: Self

4292935679

pub const AZURE: Self

4293984255

pub const LIGHTSEAGREEN: Self

4280332970

pub const TURQUOISE: Self

4283796272

pub const POWDERBLUE: Self

4289774812

pub const DARKSLATEGRAY: Self

4281522975

pub const AQUAMARINE: Self

4289379196

pub const MEDIUMSPRINGGREEN: Self

4278258842

pub const MEDIUMAQUAMARINE: Self

4288723946

pub const SPRINGGREEN: Self

4286563067

pub const MEDIUMSEAGREEN: Self

4285864305

pub const SEAGREEN: Self

4284861671

pub const LIMEGREEN: Self

4288347394

pub const DARKGREEN: Self

4284827999

pub const GREEN: Self

4284829695

pub const LIME: Self

4286563010

pub const FORESTGREEN: Self

4286526978

pub const DARKSEAGREEN: Self

4292029967

pub const LIGHTGREEN: Self

4292148896

pub const PALEGREEN: Self

4292430344

pub const MINTCREAM: Self

4294966780

pub const HONEYDEW: Self

4293987744

pub const CHARTREUSE: Self

4289370336

pub const LAWNGREEN: Self

4289336832

pub const OLIVEDRAB: Self

4286085249

pub const DARKOLIVEGREEN: Self

4283676211

pub const YELLOWGREEN: Self

4292272306

pub const GREENYELLOW: Self

4293960297

pub const BEIGE: Self

4293787924

pub const LINEN: Self

4294635702

pub const LIGHTGOLDENRODYELLOW: Self

4294637750

pub const OLIVE: Self

4286595072

pub const YELLOW: Self

4294967040

pub const LIGHTYELLOW: Self

4294967216

pub const IVORY: Self

4294967280

pub const DARKKHAKI: Self

4290623339

pub const KHAKI: Self

4293988284

pub const PALEGOLDENROD: Self

4293467754

pub const WHEAT: Self

4293720580

pub const GOLD: Self

4294956800

pub const LEMONCHIFFON: Self

4294965453

pub const PAPAYAWHIP: Self

4294955917

pub const DARKGOLDENROD: Self

4286903435

pub const GOLDENROD: Self

4292519200

pub const ANTIQUEWHITE: Self

4294634455

pub const CORNSILK: Self

4294966116

pub const OLDLACE: Self

4294960102

pub const MOCCASIN: Self

4294953679

pub const NAVAJOWHITE: Self

4294956353

pub const ORANGE: Self

4294944000

pub const BISQUE: Self

4294953924

pub const TAN: Self

4292907392

pub const DARKORANGE: Self

4294940672

pub const BURLYWOOD: Self

4292783287

pub const SADDLEBROWN: Self

4287317267

pub const SANDYBROWN: Self

4294096452

pub const BLANCHEDALMOND: Self

4294962125

pub const LAVENDERBLUSH: Self

4294964715

pub const SEASHELL: Self

4294965998

pub const FLORALWHITE: Self

4294966496

pub const SNOW: Self

4294967182

pub const PERU: Self

4286078079

pub const PEACHPUFF: Self

4294956489

pub const CHOCOLATE: Self

4289864222

pub const SIENNA: Self

4288696877

pub const LIGHTSALMON: Self

4294942842

pub const CORAL: Self

4294934352

pub const DARKSALMON: Self

4293997764

pub const MISTYROSE: Self

4294954237

pub const ORANGERED: Self

4294940672

pub const SALMON: Self

4294947974

pub const TOMATO: Self

4294943895

pub const ROSYBROWN: Self

4292080495

pub const PINK: Self

4294961403

pub const INDIANRED: Self

4289864222

pub const LIGHTCORAL: Self

4293959497

pub const BROWN: Self

4289014314

pub const FIREBRICK: Self

4289501062

pub const BLACK: Self

4278190080

pub const DIMGRAY: Self

4285098345

pub const GRAY: Self

4286611584

pub const DARKGRAY: Self

4289309097

pub const SILVER: Self

4289374890

pub const LIGHTGREY: Self

4292072403

pub const GAINSBORO: Self

4292332748

pub const WHITESMOKE: Self

4294506751

pub const WHITE: Self

4294967295

pub const GREY: Self

4287137928

pub const GREY25: Self

4282400896

pub const GREY50: Self

4286611584

pub const GREY75: Self

4290822336

impl ExtendedColors

pub const fn empty() -> Self

Get a flags value with all bits unset.

pub const fn all() -> Self

Get a flags value with all known bits set.

pub const fn bits(&self) -> u32

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u32) -> Option<Self>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u32) -> Self

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u32) -> Self

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<Self>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesnโt correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: Self) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: Self) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: Self)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: Self)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove wonโt truncate other, but the ! operator will.

pub fn toggle(&mut self, other: Self)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: Self, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: Self) -> Self

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: Self) -> Self

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: Self) -> Self

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference wonโt truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Self) -> Self

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> Self

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl ExtendedColors

pub const fn iter(&self) -> Iter<ExtendedColors>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<ExtendedColors>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.